+++ examples/anoncreds-register-revocation-registry.ts
@@ -10,6 +10,7 @@ import {
   AnonCredsSchema,
 } from '../packages/anoncreds/src/specification';
 import { PrivateKey } from '@hashgraph/sdk';
+import { Signer } from '@hiero-did-sdk/signer-internal';
 
 const operatorId = process.env.HEDERA_TESTNET_OPERATOR_ID;
 const operatorKey = process.env.HEDERA_TESTNET_OPERATOR_KEY;
@@ -25,7 +26,8 @@ const config: HederaClientConfiguration = {
 };
 
 const issuerDid = 'did:hedera:testnet:zFAeKMsqnNc2bwEsC8oqENBvGqjpGu9tpUi3VWaFEBXBo_0.0.5896419';
-const issuerKeyDer = PrivateKey.generate().toStringDer();
+const issuerKey = PrivateKey.generate();
+const issuerKeySigner = new Signer(issuerKey);
 
 const schemaPayload: AnonCredsSchema = {
   issuerId: '',
@@ -81,7 +83,7 @@ async function main() {
     console.log(`Schema registering...`);
     const schemaResult = await registry.registerSchema({
       schema: { ...schemaPayload, issuerId: issuerDid },
-      issuerKeyDer,
+      issuerKeySigner,
     });
     const schemaId = schemaResult.schemaState.schemaId;
     console.log(`Schema registered (schemaId = ${schemaId})`);
@@ -97,7 +99,7 @@ async function main() {
       options: {
         supportRevocation: true,
       },
-      issuerKeyDer,
+      issuerKeySigner,
     });
     const credDefId = credDefResult.credentialDefinitionState.credentialDefinitionId ?? '';
     console.log(`Credential definition registered (credDefId = ${credDefId})`);
@@ -110,7 +112,7 @@ async function main() {
         issuerId: issuerDid,
         credDefId,
       },
-      issuerKeyDer,
+      issuerKeySigner,
     });
     console.log('Revocation registry register result:', result);
   } catch (error) {