+++ packages/core/tests/did-message.spec.ts
@@ -1,5 +1,14 @@
 import { TestDIDMessage, TestVerifier } from './helpers';
 import { Buffer } from 'buffer';
+import { Signer } from '@hiero-did-sdk/core';
+
+const mockSignature = new Uint8Array([1, 2, 3, 4]);
+
+const mockSigner = new (class extends Signer {
+  publicKey = jest.fn();
+  sign = jest.fn().mockResolvedValue(mockSignature);
+  verify = jest.fn().mockResolvedValue(true);
+})();
 
 describe('DID Message', () => {
   let message: TestDIDMessage;
@@ -24,19 +33,18 @@ describe('DID Message', () => {
     const messagePayload = {
       foo: 'bar',
     };
-    const signature = new Uint8Array([1, 2, 3, 4]);
     const verifier = new TestVerifier();
     verifier.verifyMock.mockReturnValue(true);
 
     message.messageMock.mockReturnValue(messagePayload);
-    await message.setSignature(signature, verifier);
+    await message.setSignature(mockSignature, verifier);
 
     const payload = message.payload;
 
     expect(typeof payload).toBe('string');
     expect(JSON.parse(payload)).toMatchObject({
       message: messagePayload,
-      signature: Buffer.from(signature).toString('base64'),
+      signature: Buffer.from(mockSignature).toString('base64'),
     });
   });
 
@@ -73,36 +81,24 @@ describe('DID Message', () => {
   });
 
   it('should sign the message with the given signer', async () => {
-    const signature = new Uint8Array([1, 2, 3, 4]);
-    const signer = {
-      sign: jest.fn().mockResolvedValue(signature),
-      publicKey: jest.fn(),
-      verify: jest.fn(),
-    };
     const verifier = {
       publicKey: jest.fn(),
       verify: jest.fn().mockResolvedValue(true),
     };
 
-    await message.signWith(signer, verifier);
+    await message.signWith(mockSigner, verifier);
 
-    expect(signer.sign).toHaveBeenCalledWith(message.messageBytes);
-    expect(message.signature).toEqual(signature);
+    expect(mockSigner.sign).toHaveBeenCalledWith(message.messageBytes);
+    expect(message.signature).toEqual(mockSignature);
   });
 
   it('should throw an error if the signer is invalid', async () => {
-    const signature = new Uint8Array([1, 2, 3, 4]);
-    const signer = {
-      sign: jest.fn().mockResolvedValue(signature),
-      publicKey: jest.fn(),
-      verify: jest.fn(),
-    };
     const verifier = {
       publicKey: jest.fn(),
       verify: jest.fn().mockResolvedValue(false),
     };
 
-    await expect(() => message.signWith(signer, verifier)).rejects.toThrow(
+    await expect(() => message.signWith(mockSigner, verifier)).rejects.toThrow(
       'The signature is invalid. Provided signer does not match the DID signer.'
     );