+++ packages/messages/tests/unit/did-add-service/default-lifecycle.spec.ts
@@ -1,23 +1,25 @@
-import {
-  LifecycleRunner,
-  RunnerState,
-} from '@hiero-did-sdk/lifecycle';
+import { LifecycleRunner, RunnerState } from '@hiero-did-sdk/lifecycle';
 import { TopicMessageSubmitTransaction } from '@hashgraph/sdk';
-import {
-  DIDAddServiceMessage,
-  DIDAddServiceMessageHederaDefaultLifeCycle,
-} from '../../../src';
-import {
-  SIGNATURE,
-  TestVerifier,
-  VALID_DID,
-  VALID_DID_TOPIC_ID,
-} from '../helpers';
+import { DIDAddServiceMessage, DIDAddServiceMessageHederaDefaultLifeCycle } from '../../../src';
+import { SIGNATURE, TestVerifier, VALID_DID, VALID_DID_TOPIC_ID } from '../helpers';
+import { Signer } from '@hiero-did-sdk/core';
+
+const mockSigner = new (class extends Signer {
+  publicKey = jest.fn();
+  sign = jest.fn().mockImplementation(() => SIGNATURE);
+  verify = jest.fn().mockResolvedValue(true);
+})();
+
+const mockPublisher = {
+  network: jest.fn(),
+  publicKey: jest.fn(),
+  publish: jest.fn().mockResolvedValue({
+    topicId: VALID_DID_TOPIC_ID,
+  }),
+};
 
 describe('Default DIDAddServiceMessage Lifecycle', () => {
   describe('when processing a valid DIDAddServiceMessage', () => {
-    let publishMock: jest.Mock;
-    let signMock: jest.Mock;
     let message: DIDAddServiceMessage;
     let result: RunnerState<DIDAddServiceMessage>;
 
@@ -32,28 +34,10 @@ describe('Default DIDAddServiceMessage Lifecycle', () => {
 
       verifier.verifyMock.mockResolvedValue(true);
 
-      publishMock = jest.fn().mockResolvedValue({
-        topicId: VALID_DID_TOPIC_ID,
-      });
-
-      signMock = jest.fn().mockImplementation(() => {
-        return SIGNATURE;
-      });
-
-      const runner = new LifecycleRunner(
-        DIDAddServiceMessageHederaDefaultLifeCycle,
-      );
+      const runner = new LifecycleRunner(DIDAddServiceMessageHederaDefaultLifeCycle);
       result = await runner.process(message, {
-        signer: {
-          publicKey: jest.fn(),
-          sign: signMock,
-          verify: jest.fn(),
-        },
-        publisher: {
-          network: jest.fn(),
-          publicKey: jest.fn(),
-          publish: publishMock,
-        },
+        signer: mockSigner,
+        publisher: mockPublisher,
         args: {
           verifier,
         },
@@ -61,7 +45,7 @@ describe('Default DIDAddServiceMessage Lifecycle', () => {
     });
 
     it('should sign with given signer', () => {
-      expect(signMock).toHaveBeenCalledTimes(1);
+      expect(mockSigner.sign).toHaveBeenCalledTimes(1);
       expect(result.message.signature).toBe(SIGNATURE);
     });
 
@@ -71,27 +55,15 @@ describe('Default DIDAddServiceMessage Lifecycle', () => {
 
     describe('when resuming the lifecycle', () => {
       beforeEach(async () => {
-        const runner = new LifecycleRunner(
-          DIDAddServiceMessageHederaDefaultLifeCycle,
-        );
+        const runner = new LifecycleRunner(DIDAddServiceMessageHederaDefaultLifeCycle);
         result = await runner.resume(result, {
-          signer: {
-            publicKey: jest.fn(),
-            sign: signMock,
-            verify: jest.fn(),
-          },
-          publisher: {
-            network: jest.fn(),
-            publicKey: jest.fn(),
-            publish: publishMock,
-          },
+          signer: mockSigner,
+          publisher: mockPublisher,
         });
       });
 
       it('should publish the message to the topic', () => {
-        expect(publishMock).toHaveBeenCalledWith(
-          expect.any(TopicMessageSubmitTransaction),
-        );
+        expect(mockPublisher.publish).toHaveBeenCalledWith(expect.any(TopicMessageSubmitTransaction));
       });
     });